refactor: structured with consistent folder structure
