+++ src/main.ts
@@ -2,15 +2,20 @@ import { ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import * as morgan from 'morgan';
 import { AppModule } from './app.module';
+import { HttpExceptionFilter } from './filters/http-exception.filter';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
+  // pipes
   app.useGlobalPipes(
     new ValidationPipe({
       whitelist: true,
-      stopAtFirstError:true
-    }),
+      stopAtFirstError: true,
+    })
   );
+
+  // filters
+  app.useGlobalFilters(new HttpExceptionFilter());
 
   // morgan
   app.use(morgan('dev'));
