+++ src/resources/auth/strategy/accessToken.strategy.ts
