+++ src/resources/auth/strategy/index.ts
