+++ src/resources/bookmark/bookmark.module.ts
