+++ src/resources/bookmark/bookmark.service.ts
