+++ src/resources/bookmark/dto/update-bookmark.dto.ts
