+++ src/resources/bookmark/entities/bookmark.entity.ts
